MM-1158 extract Omnigent artifact publisher
